worked on debugging and added socket
