+++ controllers/testSocket.js
@@ -1,0 +1,7 @@
+const router = require('express').Router()
+
+router.get('/', (req, res) =>{
+    res.render('test')
+})
+
+module.exports = router;
