+++ package.json
@@ -6,7 +6,8 @@
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1",
         "start": "node server.js",
-        "seed": "node seeds/seed.js"
+        "seed": "node seeds/seed.js",
+        "dev": "nodemon server.js"
     },
     "keywords": [],
     "author": "Conor, Greg, Whitney, Heinz, Ernest",
