+++ seeds/userData.json
@@ -1,8 +1,8 @@
-{
-  "users": [
+ [
     {
-      "userId": 1,
-      "username": "WhimsyMaster",
+      "first_name": "JJJJ",
+      "last_name": "JJJ",
+      "user_name": "WhimsyMaster",
       "password": "SillyPass123",
       "email": "whimsymaster@example.com",
       "imageUrl": "public/images/pP3.jpeg",
@@ -11,8 +11,9 @@
       "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
     },
     {
-      "userId": 2,
-      "username": "LaughingLuna",
+      "first_name": "LLL",
+      "last_name": "LLL",
+      "user_name": "LaughingLuna",
       "password": "Giggles456",
       "email": "laughingluna@gmail.com",
       "imageUrl": "public/images/pP6.jpeg",
@@ -21,8 +22,9 @@
       "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
     },
     {
-      "userId": 3,
-      "username": "QuirkQuester",
+      "first_name": "PPP",
+      "last_name": "PPP",
+      "user_name": "QuirkQuester",
       "password": "JovialPass789",
       "email": "quirkquester@hotmail.com",
       "imageUrl": "public/images/pP9.jpeg",
@@ -31,74 +33,15 @@
       "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
     },
     {
-      "userId": 4,
-      "username": "ChuckleChampion",
+      "first_name": "OOO",
+      "last_name": "OOO",
+      "user_name": "ChuckleChampion",
       "password": "WackyCode987",
       "email": "chucklechampion@yahoo.com",
       "imageUrl": "public/images/pP12.jpeg",
       "posts": [7, 12],
       "comments": [3, 12, 22, 33],
       "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
-    },
-    {
-      "userId": 5,
-      "username": "JestJester",
-      "password": "Zany123Pass",
-      "email": "jestjester@outlook.com",
-      "imageUrl": "public/images/pP15.jpeg",
-      "posts": [3, 18],
-      "comments": [4, 13, 23, 34],
-      "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
-    },
-    {
-      "userId": 6,
-      "username": "MirthfulMingle",
-      "password": "Joyful456Code",
-      "email": "mirthfulmingle@gmail.com",
-      "imageUrl": "public/images/pP18.jpeg",
-      "posts": [8, 13],
-      "comments": [5, 14, 24, 35],
-      "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
-    },
-    {
-      "userId": 7,
-      "username": "GuffawGuru",
-      "password": "SnickerPass789",
-      "email": "guffawguru@example.com",
-      "imageUrl": "public/images/pP21.jpeg",
-      "posts": [4, 19],
-      "comments": [6, 15, 25, 36],
-      "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
-    },
-    {
-      "userId": 8,
-      "username": "WhoopeeWizard",
-      "password": "Hilarious987Key",
-      "email": "whoopeewizard@yahoo.com",
-      "imageUrl": "public/images/pP24.jpeg",
-      "posts": [9, 14],
-      "comments": [7, 16, 26, 37],
-      "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
-    },
-    {
-      "userId": 9,
-      "username": "GleefulGiggler",
-      "password": "FunTime123Pass",
-      "email": "gleefulgiggler@hotmail.com",
-      "imageUrl": "public/images/pP25.jpeg",
-      "posts": [5, 20],
-      "comments": [8, 17, 27, 38],
-      "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
-    },
-    {
-      "userId": 10,
-      "username": "SnickerSage",
-      "password": "JollyPass456",
-      "email": "snickersage@outlook.com",
-      "imageUrl": "public/images/pP28.jpeg",
-      "posts": [10, 15],
-      "comments": [9, 18, 28, 39],
-      "post_rating": [4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3, 4, 5, 3]
     }
   ]
-}
+
