+++ server.js
@@ -5,10 +5,19 @@ const exphbs = require('express-handlebars');
 const routes = require('./controllers');
 const helpers = require('./utils/helpers');
 
+//socketio
+const {createServer} = require('node:http');
+const {fileURLToPath} = require('node:url');
+const {dirname, join } = require('node:path');
+const {Server} = require('socket.io')
+
 const sequelize = require('./config/connection');
 const SequelizeStore = require('connect-session-sequelize')(session.Store);
 
 const app = express();
+const server = createServer(app);
+const io = new Server(server);
+// const projectDirectory = dirname(fileURLToPath(require(meta.url)));
 const PORT = process.env.PORT || 3001;
 
 // Set up Handlebars.js engine with custom helpers
@@ -38,14 +47,25 @@ app.set('view engine', 'handlebars');
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(join(__dirname, 'public')));
 
 app.use(routes);
 
 // Serve static files from the 'public' folder (Heinz Add-on)
-app.use('/images', express.static(path.join(__dirname, 'public/images')));
+app.use('/images', express.static(join(__dirname, 'public/images')));
+
+//sends message to everyone including sender
+io.on('connection', (socket) => {
+  socket.on('chat message', (msg) => {
+    io.emit('chat message', msg);
+  });
+  
+});
 
 
 sequelize.sync({ force: false }).then(() => {
-  app.listen(PORT, () => console.log('http://localhost:3001'));
+  server.listen(PORT, () => {
+    console.log(`server running at http://localhost:${PORT}`);
+    });
+  // app.listen(PORT, () => console.log('http://localhost:3001'));
 });
