+++ views/homepage.handlebars
@@ -10,8 +10,9 @@
   </div>
   <div class="col-md-7">
     <p>
-      {{post.content}}
+      {{post.description}}
     </p>
   </div>
 </div>
 {{/each}}
+{{> test}}
